Fix all 25 bugs found by the full code scan
Renderer/scenes: clear uniform cache on shader recompile (uniforms silently
corrupted after editor recompiles); integrate shader time (speed changes/LFO
no longer scrub uTime); NEAREST audio texture (R32F+LINEAR incomplete
without OES_texture_float_linear); outgoing scene keeps FX during FADE/MELT
transitions; milk preset load/reload applied on the GL thread with a
synchronized queue poll; delete ProjectMScene post VAO; triangle-wave
endless-zoom exponent in all 20 scene shaders (wrap pop); LFO S&H phase in
double precision.

Audio/analysis: bound AIFF PCM reads to the SSND payload (trailing chunks
decoded as noise); unsigned chunk sizes + robust skips + odd-bit-depth
rejection in AiffPcm; COMM validation, typed ParserExceptions and 64-bit
skips in AiffExtractor; OfflineAnalyzer releases extractor/codec on all
failure paths; Pearson (mean-centered) key correlation removes the minor
bias.

Export: single try/finally covers setup AND the loop so cancel/failure
never leaks the hardware codec, surface, muxer or EGL context; encoder
re-created from the decoder's real output format (HE-AAC/SBR, parametric
stereo); rate-derived AAC tail pad; absolute-byte AIFF PTS (no hour-scale
drift); pending-export request survives rotation and orphan files/pending
MediaStore rows are cleaned up; EOS-flush window tripled and truncation
logged.

Stores/ViewModel: analysis cached under the analyzed track (and the skipped
track re-queued); synchronized + atomic TrackLibrary writes; playlist and
preset files resolved by stored JSON name so sanitize collisions can't
destroy data; MediaStore scan and queue building off the main thread;
async history writes; guarded takePersistableUriPermission (AppShell +
ExportHost); search-screen preset apply pushes the custom shader.

New tests kept green by design; versionCode 14 / versionName 0.12.1.

+++ musicviz-project/musicviz/app/src/main/java/dev/musicviz/render/Lfo.kt
@@ -85,7 +85,10 @@ class LfoEngine {
 
     private val phases = FloatArray(3)
     private val sampleHold = FloatArray(3)
-    private val totalPhase = FloatArray(3)
+
+    /** Double: float ulps exceed the per-frame increment after ~2^24 cycles,
+     *  freezing the S&H waveform on very long sessions. */
+    private val totalPhase = DoubleArray(3)
     private val lastCycle = IntArray(3) { -1 }
 
     /** Advances phases and returns each LFO's bipolar output value. */
